NOTFORMERGE: purism/librem13v2: Workaround for SATA 6gbps issues
Planet alignment and moon phases seem to affect whether or not
booting from M.2 SSD (and apparently even 2.5" HDD) but original
vendor is not affected by it. The only thing I could find is that
the ATA errors we see on linux in dmesg disappear after linux switches
to a 3Gbps speed limit on the controller.

This patch will force the ATA ports 0 and 2 to 3Gbps speed limit which
will allow grub to read from the discs without problems. Once linux
boots, it resets the controller which removes this hack, and linux
will more appropriately drop back the speed to 3Gbps if it feels the
need for it.

ATA spec says wait for 1ms, but there's no guarantee that the reset
will be done after 1ms, so I wait 5ms instead. It's bad, it's ugly,
but this is a temporary hack anyway.

Change-Id: If785c63d92a19044554828798d31e450bcd78d87
Signed-off-by: Youness Alaoui <youness.alaoui@puri.sm>