soc/intel: Remove chipset_lockdown configuration option
For Skylake/Kabylake/Apollolake/Coffeelake, the chipset lockdown option
between FSP and COREBOOT is not useful as everyone sets it to COREBOOT
and the FSP path is broken anyway (According to Nico Huber).
This removes the config option and always assumes the path of locking
down the chipset by coreboot.

Change-Id: I9dbbb5e409e09fa8a73c1a3f9b1f8383a4498c58
Signed-off-by: Youness Alaoui <youness.alaoui@puri.sm>